Fully test collideV method, minor refactoring
Added positive and negative test cases that hinge on each boolean
posibility while exercising boundry conditions.

Seperated command from query on individual platform collission, and
renamed function to willlFallOn(platform) to convey meaning

Introduced loop flag named collisionDetected to decide whether to fall, so
i can avoid checking the status of i to detect a break which is bad practice

Moved changeMovieClipHeight and changeMovieClipWidth to Util class, as it
is nopn project specific and is at the wrong level of abstraction for the
test class.