Move some test lists from suites into test classes
It makes more sense to have each TestCase provide it's own list of test
functions rather than having an external class do this.

I moved the TestSuite construction logic into a function named getSuite
on the TestCases that build up a TestSuite object using referances to
the functions from whithin the class. I use a class named
ReflectionTestSuiteBuilder get the name of the function from a referance
to the function in order to decrease the reliance on direct reflection.

So far i have done this to the following classes:

@ ColliderTest
@ KeyboardControlsTest
@ UserInterfaceManagerTest

The rest of the tests are to be commited soon.